Refactor code, deal with norm
